Actually make pycld3 optional: allow it be missing in imports and unit tests
